+++ src/components/AccountTypePicker/AccountTypePicker.jsx
@@ -1,11 +1,4 @@
-import {
-  Box,
-  FormControl,
-  InputLabel,
-  MenuItem,
-  Select,
-  Typography,
-} from '@mui/material';
+import { Box, InputLabel, MenuItem, Select } from '@mui/material';
 import PropTypes from 'prop-types';
 import { useState } from 'react';
 import { useDispatch } from 'react-redux';
@@ -30,26 +23,24 @@ export default function AccountTypePicker({ account }) {
 
   return (
     <Box>
-      <Typography variant='h5'>Account Type</Typography>
-      <FormControl sx={{ m: 1, minWidth: 120 }}>
-        <InputLabel id='account-type-label'>Account Type</InputLabel>
-        <Select
-          labelId='account-type-label'
-          id='account-type'
-          value={type}
-          label='Account Type'
-          onChange={handleChange}
-        >
-          {Object.keys(constants.AccountType).map((key) => (
-            <MenuItem
-              key={key}
-              value={constants.AccountType[key]}
-            >
-              {constants.AccountType[key]}
-            </MenuItem>
-          ))}
-        </Select>
-      </FormControl>
+      <InputLabel id='account-type-label'>Account Type</InputLabel>
+      <Select
+        labelId='account-type-label'
+        id='account-type'
+        value={type}
+        label='Account Type'
+        onChange={handleChange}
+        style={{ width: '150px' }}
+      >
+        {Object.keys(constants.AccountType).map((key) => (
+          <MenuItem
+            key={key}
+            value={constants.AccountType[key]}
+          >
+            {constants.AccountType[key]}
+          </MenuItem>
+        ))}
+      </Select>
     </Box>
   );
 }
