+++ src/components/BalanceDisplay/index.js
@@ -1,0 +1,1 @@
+export { default } from './BalanceDisplay';
