+++ src/components/SettingsPanel/SettingsPanelItem.jsx
@@ -1,0 +1,10 @@
+import styled from '@emotion/styled';
+import { Box } from '@mui/material';
+
+export const SettingsPanelItem = styled(Box)`
+  width: 200px;
+  margin-top: 20px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+`;
